Disallow private-use characters used internally.
A set of 16 private-use characters are used internally to aid in parsing and
simplify the code.  If these characters were not used, parsing would be much,
much slower and more complex.  This adjusts the main troff code to raise an
exception if one of these characters is found on input.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>